[Patch] Modified the CLI create command
